git-svn-id: https://svn.apache.org/repos/asf/incubator/jena/Import/Jena-SVN/ARQ/trunk@1120633 13f79535-47bb-0310-9956-ffa450edef68

+++ src-test/com/hp/hpl/jena/sparql/suites/optimizer/TestGraphBuild.java
@@ -53,8 +53,8 @@ public class TestGraphBuild extends TestCase
 	    
 	    for (int i = 0; i < components.size(); i++)
 	    {	    	
-	    	Integer nodes = (Integer)components.get(i).get("nodes") ;
-	    	Integer edges = (Integer)components.get(i).get("edges") ;
+	    	Integer nodes = components.get(i).get("nodes") ;
+	    	Integer edges = components.get(i).get("edges") ;
 	    	ConnectedGraph component = graph.getComponent(i) ;
 	    	
 	    	assertTrue(component.getNodes().size() == nodes.intValue()) ;
